SECCOMP Changes
The original operating mode of Enforcing placed a policy on the entire sandbox. This meant that bubblewrap itself was confined, which was good for security, but also brought in all the syscalls bubblewrap needed into the application sandbox. To remedy this, Antimony now creates a BPF for bubblewrap to use itself, such that bubblewrap is confined under one filter, and then the sandbox is confined under a more restrictive filter, removing the bubblewrap syscalls.

Also added a new SECCOMP mode: Notify. By using notify-send, we can directly prompt the user when unrecognized syscalls occur in the sandbox. Users can then Allow for one-time, Deny for the remainder of the session, Kill the entire process group with SIGTERM, or Save the syscall in the same way Permissive worked. This provides an excellent default, where after Permissive has collected the majority of syscalls, Notify can be used to prompt the user if a new one shows up--rather than in enforcing where the entire sandbox is killed without giving the opportunity to realize what syscall was attempted.

+++ src/cli/edit.rs
@@ -1,5 +1,5 @@
 //! Edit an existing profile.
-use crate::aux::profile::Profile;
+use crate::{aux::profile::Profile, cli::default};
 use anyhow::Result;
 use std::fs;
 
@@ -10,26 +10,31 @@ pub struct Args {
 }
 impl super::Run for Args {
     fn run(self) -> Result<()> {
-        // Edit the profile
-        let user = Profile::user_profile(&self.profile);
-        let new = !user.exists();
-        if new {
-            let source = Profile::path(&self.profile)?;
-            if let Some(parent) = user.parent() {
-                user::set(user::Mode::Effective)?;
-                fs::create_dir_all(parent)?;
-                user::revert()?;
+        if self.profile == "default" {
+            let args = default::Args::default();
+            args.run()
+        } else {
+            // Edit the profile
+            let user = Profile::user_profile(&self.profile);
+            let new = !user.exists();
+            if new {
+                let source = Profile::path(&self.profile)?;
+                if let Some(parent) = user.parent() {
+                    user::set(user::Mode::Effective)?;
+                    fs::create_dir_all(parent)?;
+                    user::revert()?;
+                }
+                fs::copy(source, &user)?;
             }
-            fs::copy(source, &user)?;
-        }
 
-        // Edit it.
-        if Profile::edit(&user)?.is_none() && new {
-            // If there was no modifications, delete the profile
-            // since it's identical to the system one.
-            user::set(user::Mode::Effective)?;
-            fs::remove_file(user)?;
+            // Edit it.
+            if Profile::edit(&user)?.is_none() && new {
+                // If there was no modifications, delete the profile
+                // since it's identical to the system one.
+                user::set(user::Mode::Effective)?;
+                fs::remove_file(user)?;
+            }
+            Ok(())
         }
-        Ok(())
     }
 }
